requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
idna==3.4
kombu==5.3.2
magic-filter==1.0.11
multidict==6.0.4
outcome==1.2.0
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
